New Container Class
I going to delete all the classes with the chips. /n
Because i'm found a better way to make the behavior of them /n
File: chess_chips.py
By: Nelcido
Date: 23/4:35 Am
